models/ddpm/requirements.txt
accelerate==0.20.3
aiohttp==3.8.4
aiosignal==1.3.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
bcrypt==4.0.1
beartype==0.14.1
bleach==6.0.0
bokeh==3.2.0
boto3==1.26.163
botocore==1.29.163
braceexpand==0.1.7
cachetools==5.3.1
certifi==2023.5.7
click==8.1.3
clip-anytorch==2.5.2
cmake==3.26.4
CoCa-pytorch==0.0.7
colorcet==3.0.1
contextlib2==21.6.0
contourpy==1.1.0
dalle2-pytorch==1.14.2
datasets==2.13.1
dill==0.3.6
einops==0.6.1
einops-exts==0.0.4
ema-pytorch==0.2.3
embedding-reader==1.5.1
filelock==3.12.2
flatbuffers==23.5.26
frozenlist==1.3.3
fsspec==2023.6.0
ftfy==6.1.1
gast==0.4.0
google-auth==2.21.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.56.0
h5py==3.9.0
holoviews==1.16.2
huggingface-hub==0.15.1
imagen==2.1.0
imagen-pytorch==1.25.4
jax==0.4.13
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.1
keras==2.13.1
Keras-Preprocessing==1.1.2
kornia==0.6.12
libclang==16.0.0
lightning-utilities==0.9.0
linkify-it-py==2.0.2
lion-pytorch==0.1.2
lit==16.0.6
lpips==0.1.4
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdit-py-plugins==0.4.0
mdurl==0.1.2
ml-collections==0.1.1
ml-dtypes==0.2.0
multidict==6.0.4
multiprocess==0.70.14
nltk==3.8.1
numpy==1.23.5
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.2.1.16
nvidia-cuda-runtime-cu12==12.2.53
nvidia-cudnn-cu11==8.6.0.163
nvidia-cudnn-cu12==8.9.2.26
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
pandas==1.5.3
panel==1.2.0
param==1.13.0
paramiko==3.2.0
protobuf==4.23.3
psutil==5.9.5
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyct==0.5.0
pydantic==1.10.9
PyNaCl==1.5.0
python-dateutil==2.8.2
python-version==0.0.2
pytorch-lightning==2.0.4
pytorch-warmup==0.1.1
pytz==2023.3
pyviz-comms==2.3.2
PyYAML==6.0
regex==2023.6.3
requests-oauthlib==1.3.1
resize-right==0.0.2
rotary-embedding-torch==0.2.3
rsa==4.9
s3transfer==0.6.1
safetensors==0.3.1
scikit-learn==1.3.0
scipy==1.10.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.32.0
tensorrt==8.6.1
termcolor==2.3.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.1.0
timm==0.9.2
tokenizers==0.13.3
tornado==6.3.2
transformers==4.30.2
triton==2.0.0.post1
typing_extensions==4.5.0
uc-micro-py==1.0.2
vector-quantize-pytorch==1.6.24
wcwidth==0.2.6
webdataset==0.2.48
webencodings==0.5.1
Werkzeug==2.3.6
wrapt==1.14.1
x-clip==0.12.1
xxhash==3.2.0
xyzservices==2023.5.0
yarl==1.9.2
zstandard==0.19.0
matplotlib
opencv-python
piq
ml-collections
